# caseframesvc/src/main/java/uk/gov/ons/ctp/response/caseframe/domain/model/Region.java
package uk.gov.ons.ctp.response.caseframe.domain.model;

import java.io.Serializable;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * Domain model object
 */
@Entity
@Data
@AllArgsConstructor
@NoArgsConstructor(access = AccessLevel.PUBLIC)
@Table(name="region", schema="refdata")
public class Region implements Serializable {

  private static final long serialVersionUID = 3552687869535168637L;

  @Id
  private String rgn11cd;

  private String rgn11nm;

}