# caseframesvc/swagger.yml
swagger: '2.0'
info:
  title: Case Frame Service API
  description: Provide access to Address Frame and Response Management information
  version: 8.11.0
schemes:
  - http
basePath: /frameservice
produces:
  - application/json
...
paths:
  /regions:
    get:
      summary: Get all regions
      description: >
        Return the Nine GSS Regional and Geography Committee Regions of England
        (formerly GOR) plus region for Wales.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: An array of regions
          schema:
            type: array
            items:
              $ref: '#/definitions/Region'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /regions/{regionid}:
    get:
      summary: Get information for the specified region.
      description: |
        Return information for the supplied Region code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: regionid
          in: path
          description: Nine character Region code, E12000001 to E12000009 or W92000004.
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: A single region
          schema:
              $ref: '#/definitions/Region'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /regions/{regionid}/lads:
    get:
      summary: Get all LADs for the specified region.
      description: |
        Return the Local Authority Districts (LADs) for the supplied Region code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: regionid
          in: path
          description: Nine character region code, E12000001 to E12000009 or W92000004.
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: An array of LADs
          schema:
            type: array
            items:
              $ref: '#/definitions/LocalAuthDist'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /lads/{ladid}:
    get:
      summary: Get information for the specified LAD.
      description: |
        Return information for the supplied Local Authority District (LAD) code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: ladid
          in: path
          description: Nine character LAD code e.g. E06000001, E09000033, W06000024.
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: A single LAD.
          schema:
              $ref: '#/definitions/LocalAuthDist'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /lads/{ladid}/msoas:
    get:
      summary: Get all MSOAs for the specified LAD.
      description: |
        Return the Mid Layer Super Output Area (MSOA) codes for the supplied Local Authority District (LAD).
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: ladid
          in: path
          description: Nine character LAD code.
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: An array of MSOAs for the supplied LAD code.
          schema:
            type: array
            items:
              $ref: '#/definitions/MSOA'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /msoas/{msoaid}:
    get:
      summary: Get information for the specified MSOA.
      description: |
        Return information for the supplied Mid Layer Super Output Area (MSOA) code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: msoaid
          in: path
          description: Nine character MSOA code.
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: A single MSOA.
          schema:
              $ref: '#/definitions/MSOA'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /msoas/{msoaid}/addresssummaries:
    get:
      summary: Get address summaries for the specified MSOA.
      description: |
        Return address summaries for the supplied Mid Layer Super Output Area (MSOA) code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: msoaid
          in: path
          description: Nine character MSOA code.
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: An array of address summaries for the supplied MSOA code.
          schema:
            type: array
            items:
              $ref: '#/definitions/AddressSummary'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /questionnaires/iac/{iac}:
    get:
      summary:  Get the questionnaire details for a specified IAC.
      description: |
        Return the questionnaire details for the supplied Internet Access Code (IAC).
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: iac
          description: Unique Internet Access Code (token).
          in: path
          required: true
          type: string
      tags:
        - Response
      responses:
        '200':
          description: Details for a single questionnaire.
          schema:
            $ref: '#/definitions/Questionnaire'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /questionnaires/case/{caseid}:
    get:
      summary:  Get all questionnaire details for a specified case.
      description: |
        Return all questionnaire details for the supplied caseid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: caseid
          description: Unique case Id.
          in: path
          required: true
          type: string
      tags:
        - Response
      responses:
        '200':
          description: All questionnaire details for a supplied caseid.
          schema:
            type: array
            items:
              $ref: '#/definitions/Questionnaire'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /questionnaires/{questionnaireid}/response:
                put:
                  summary:  Confirms a response has been received in the Survey Data Exchange.
                  description: |
                    For the given Questionnaire, it populates the response_datetime. It also populates case_status for the related Case.
                  parameters:
                    - name: Accept
                      in: header
                      description: Response type, application/json.
                      required: false
                      type: string
                    - name: questionnaireid
                      description: Unique questionnaire Id.
                      in: path
                      required: true
                      type: string
                  tags:
                    - Response
                  responses:
                    '200':
                      description: The response mechanism was completed successfully.
                    default:
                      description: Unexpected error
                      schema:
                        $ref: '#/definitions/Error'
...
  /questionsets:
    get:
      summary: Get all question set types.
      description: >
        Return all question set types and their description.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
      tags:
        - Response
      responses:
        '200':
          description: An array of question sets
          schema:
            type: array
            items:
              $ref: '#/definitions/QuestionSet'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /questionsets/{questionset}:
    get:
      summary: Get information for the specified question set.
      description: Return information for the supplied question set.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: questionset
          in: path
          description: Unique question set code.
          required: true
          type: string
      tags:
        - Response
      responses:
        '200':
          description: A single question set.
          schema:
              $ref: '#/definitions/QuestionSet'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Region:
    type: object
    properties:
      regionCode:
        type: string
        description: GSS nine character Region geography code.
      regionName:
        type: string
        description: Region name.
  LocalAuthDist:
    type: object
    properties:
      regionCode:
        type: string
        description: GSS nine character Region geography code.
      ladCode:
        type: string
        description: GSS nine character LAD code.
      ladName:
        type: string
        description: Local Authority Distict name.
  MSOA:
    type: object
    properties:
      msoaCode:
        type: string
        description: GSS nine character MSOA geography code.
      msoaName:
        type: string
        description: MSOA name.
  AddressSummary:
    type: object
    properties:
      uprn:
        type: integer
        description: Unique Property Reference Number.
      addressType:
        type: string
        description: Address type CE, HH, I.
      estabType:
        type: string
        description: Communal Establishment type.
      addressLine1:
        type: string
        description: Address details.
      addressLine2:
        type: string
        description: Address details.
      townName:
        type: string
        description: Postal town name.
      postcode:
        type: string
        description: Postal Code.
...
  Questionnaire:
    type: object
    properties:
      questionnaireId:
        type: integer
        description: Unique questionnaire ID.
      caseId:
        type: integer
        description: Unique case identifier.
      questionnaireStatus:
        type: string
        description: Questionnaire status.
      dispatchDateTime:
        type: string
        format: date-time
        description: Date and time the questionnaire was dispatched.
      responseDateTime:
        type: string
        format: date-time
        description: Date and time the questionnaire was sent for return.
      receiptDateTime:
        type: string
        format: date-time
        description: Date and time the questionnaire was recieved.
      questionSet:
        type: string
        description: Type of questionnaire set.
      iac:
        type: string
        description: Internet Access Code.
...
  QuestionSet:
    type: object
    properties:
      questionSet:
        type: string
        description: Type of question set.
      description:
        type: string
        description: Description of question set.
...
  Error:
    type: object
    properties:
      code:
        type: string
      timestamp:
        type: string
      message:
        type: string